Move test to test_other
